main: scaffolding works again

+++ TheCardEditor.DataModel/DataModel/Card.cs
@@ -6,9 +6,9 @@ public partial class Card
     {
     }
 
-    public long Id { get; set; }
+    public int Id { get; set; }
     public string Name { get; set; } = null!;
-    public long CardSetFk { get; set; }
+    public int CardSetFk { get; set; }
     public string Data { get; set; } = null!;
     public virtual CardSet CardSetFkNavigation { get; set; } = null!;
 }
