+++ TheCardEditor.DataModel/DataModel/Template.cs
@@ -6,9 +6,9 @@ public partial class Template
     {
     }
 
-    public long Id { get; set; }
+    public int Id { get; set; }
     public string Name { get; set; } = null!;
     public string Data { get; set; } = null!;
-    public long CardSetFk { get; set; }
+    public int CardSetFk { get; set; }
     public virtual CardSet CardSetFkNavigation { get; set; } = null!;
 }
